requirements.txt
fastapi==0.115.12
langfuse==2.60.7
pydantic==2.11.4
uvicorn==0.34.2
azure-identity==1.21.0
azure-mgmt-postgresqlflexibleservers==1.1.0
mcp==1.10.0
semantic-kernel 
python-dotenv
psycopg[binary]
ipykernel
openai
scikit-learn
numpy
pgvector
